fix(section): make sure coordinator of a section controller is nilled properly
This fix ensures that the section coordinator is only nilled on section controllers that are not needed anymore. This fixes an error that occurred when setting a new array of section controllers that partly matches the old section controller array. In this case, the old section controllers were not linked with the section coordinator anymore.