fix(section controller): ensure that there are no retain cycles in the case a section controller retains its table view
This ensures that every section controller gets an update to `tableViewDidChange:` with a `nil` table view to remove a cached table view. This prevents retain cycles between the table view and the section controller that might occur because the table view actually is a proxy that needs to retain its table view and its section controller.